Deleted assignment separation card and reorganized
