--- src/Components/Assignments/AssignmentSeparationCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './AssignmentSeparationCard';
