+++ src/Components/Assignments/Assignments.jsx
@@ -9,8 +9,8 @@ import { altAssignmentFetchData } from 'actions/assignment';
 import Spinner from 'Components/Spinner';
 import Alert from 'Components/Alert';
 import TabbedCard from 'Components/TabbedCard';
-import Assignment from './AssignmentSeparationCard/Assignment';
-import Separation from './AssignmentSeparationCard/Separation';
+import Assignment from './Assignment';
+import Separation from './Separation';
 import ReactModal from '../ReactModal';
 import InteractiveElement from '../InteractiveElement';
 import api from '../../api';
